+++ src/components/Questionnaire/index.js
@@ -1,7 +1,6 @@
 import React, { useState } from 'react'
 import { Model } from 'survey-core'
 import { Survey } from 'survey-react-ui'
-import Fuse from 'fuse.js'
 import Modal from 'react-modal'
 import 'survey-core/defaultV2.min.css'
 import './index.scss'
@@ -42,152 +41,17 @@ const Questionnaire = ({ setShowButtons }) => {
         'Master thesis',
     ]
 
-    const programsAndIndustries = [
-        {
-            program: 'Biomedical Engineering',
-            industries: ['Pharmaceutical', 'Biotechnology'],
-        },
-        {
-            program: 'Chemical Engineering',
-            industries: [
-                'Solid Mechanics',
-                'Pharmaceutical',
-                'Biotechnology',
-                'Nuclear Power',
-                'Energy Technology',
-                'Nanotechnology',
-            ],
-        },
-        {
-            program: 'Civil Engineering',
-            industries: [
-                'Architecture',
-                'Construction',
-                'Property & Infrastructure',
-                'Railway',
-            ],
-        },
-        {
-            program: 'Computer Science & Engineering',
-            industries: [
-                'Simulation Technology',
-                'Software Development',
-                'Web Development',
-                'Telecommunication',
-                'IT Infrastructure',
-                'Interaction Design',
-            ],
-        },
-        {
-            program: 'Electrical Engineering',
-            industries: [
-                'Acoustics',
-                'Aerospace',
-                'Telecommunication',
-                'Mechatronics',
-                'Electronics',
-                'Nanotechnology',
-            ],
-        },
-        {
-            program: 'Engineering Mathematics & Physics',
-            industries: [
-                'Solid Mechanics',
-                'Acoustics',
-                'Nuclear Power',
-                'Fluid Mechanics',
-                'Industry Design',
-            ],
-        },
-        {
-            program: 'Environmental Engineering',
-            industries: ['Environmental Sector', 'Energy Technology'],
-        },
-        {
-            program: 'Industrial Engineering',
-            industries: [
-                'Manufacturing Industry',
-                'Management Consulting',
-                'Insurance',
-                'Finance',
-                'Logistics & Supply Chain',
-                'Mechatronics',
-                'Property & Infrastructure',
-                'Industry Design',
-                'Recruitment',
-            ],
-        },
-        {
-            program: 'Information Technology',
-            industries: [
-                'Web Development',
-                'Simulation Technology',
-                'Telecommunication',
-                'IT Infrastructure',
-                'Software Development',
-                'Interaction Design',
-            ],
-        },
-        {
-            program: 'Mechanical Engineering',
-            industries: [
-                'Automotive',
-                'Fluid Mechanics',
-                'Solid Mechanics',
-                'Aerospace',
-                'Acoustics',
-                'Marine System',
-                'Mechatronics',
-                'Nanotechnology',
-            ],
-        },
-        {
-            program: 'Media Technology',
-            industries: [
-                'Web Development',
-                'Simulation Technology',
-                'Media Technology',
-                'Telecommunication',
-                'IT Infrastructure',
-                'Software Development',
-                'Interaction Design',
-            ],
-        },
-        {
-            program: 'Medical Engineering',
-            industries: ['Medical Technology'],
-        },
-        {
-            program: 'Material & Product Design',
-            industries: [
-                'Pharmaceutical',
-                'Wood-Processing Industry',
-                'Steel Industry',
-                'Manufacturing Industry',
-                'Logistics & Supply Chain',
-                'Material Development',
-                'Nanotechnology',
-                'Product Development',
-            ],
-        },
-        {
-            program: 'Other',
-            industries: ['Research', 'Pedagogy', 'Retail'],
-        },
-        // Add more program-industry associations
-    ]
+    // const options = {
+    //     keys: ['program'], // Search based on the program name
+    //     threshold: 0.1, // Adjust the threshold for fuzzy matching
+    // }
 
-    const options = {
-        keys: ['program'], // Search based on the program name
-        threshold: 0.1, // Adjust the threshold for fuzzy matching
-    }
+    // const fuse = new Fuse(programsAndIndustries, options) //instance for fuzzy searching
 
-    const fuse = new Fuse(programsAndIndustries, options) //instance for fuzzy searching
-
-    function matchedIndustries(program) {
-        const results = fuse.search(program)
-        return results.map(result => result.item.industries).flat()
-    }
+    // function matchedIndustries(program) {
+    //     const results = fuse.search(program)
+    //     return results.map(result => result.item.industries).flat()
+    // }
 
     const surveyJson = {
         pages: [
